Add parsing of FloatXX types
The reason I'm pasting the macros is because I got into macro hell
otherwise: a macro calling a macro calling a macro... and somewhere some
type conversion went wrong...

Currently I run the macro with cpp from docker to generate the
functions. The functions are only one line, so I process them with vim,
and then paste them in the file.

In the future I want to automate it in the build process.