Add initial parsing of Dimension element
I found memory management (for dimension array) more robust if we
handled it instead of user. The dimension array could be optimized to be
computed only once, but feels unnecessary right now.

During runtime the size of the dimension (but not the amount of
dimensions) can change, so we probably need another API call that lets
the user get the runtime dimensions sizes, or a list of {dim_idx + valueRef + is_structuralParam} structs, so they can check it themselves.