Remove or rename all fmi3Real to fmi3Float64, update some signatures
Starting with testing of float64. BouncingBall fmu must be
parametrizable for both float32 and float64 for testing float32 to work.

One refactor that must be done is to change how set/get _real works,
because currently the states and variables are float64, but it should be
possible to have float32 variables (it appears that states must be
float64). The getter/setter does not support this.

Some function signatures have also been updated for 3.0, because I just
copied from the fmi3*Types.h files.