Add tests for real attributes and refactor name
I couldn't read the code when every second word was type, base or a mix
of that: baseType, typeBase. This is the first step to make it readable.
Sometimes type_base just means a pointer to any kind of type_base, and
sometimes it means the super type. Now that part is clear at least.

Tests are failing as well, because there is an old bug.