Check that referenced units exist in UnitDefs
I didn't understand why we would need dynamic creation of new units,
except for when the FMU had invalid XML and we wanted to still be able
to run without "errors" (if the model needs to be converted with respect
to base units, the model is invalid though), so removed it. I can revert
this if found too strict.

fmil_test.c is not the best name for the test util, but I think I've
already made this extraction (of a common test util) for 3.0, with this
name.

The input argument to the test function is now the 'parser_test_xmls'
directory instead of the specific sub directory. This will require that
some paths are duplicated in the test code, but it makes it easier to
debug tests from e.g. Visual Studio, since you can always give the same
path as argument then, and don't have to modify for each test.