Find DisplayUnit via Unit instead of global vector
We stored all displayUnitDefs in a separate vector, and used bsearch on it. The
problem is that unitDisplays are not globally unique with respect to name,
which means that we might return the wrong one, since they were stored in
jm_named_ptr.

The Unit/DisplayUnit API is being tested on a very high level (print if
possible) in fmi2_import_xml_test, but not asserted with expected
compared to actual. To verify my changes I therefore decided to add some
more thorough unit tests.

Spec. says "if displayUnit is defined, but not unit, then this is invalid", but it
says nothing if a variable inherits a type with a unit. Is it then valid for the
variable to have a displayUnit (assuming it matches inherited unit)?
Currently I don't have any tests for this.
Example:
Unit: "K"
    DisplayUnit: "degC"
    DisplayUnit: "degF"
Type: "type"
    unit: "K"
    displayUnit: "degC"
Variable: "var"
    declaredType: "type"
    displayUnit: "degF"