Change test locale and add debug prints
Summary: I suspect that a test node had the Swedish locale modified.

Changed from Swedish to French because on one Windows Jenkins node the
decimal point was "point" even after successfully setting the locale
(expected to be "comma").

I couldn't find any information by searching, but after trial and error
I found that changing the locale's decimal point via the Windows
Settings GUI (on Windows 10 dev machine) were persistent, i.e. the
setting was propagated to the test.

I only managed to replicate it for 'sv-SE', and not
'Swedish_Sweden.1252' (which had been failing) via the GUI. The nodes
are Windows Server, but I don't expect much difference.

So, I suspect that a test node had the Swedish locale modified.

I don't know if the "C" locale can be modified (it's not accessible via
GUI at least), but if we want a truly robust implementation we would
have to read the localeconv()'s decimal_point field and adapt
accordingly everywhere necessary. For example, I believe the
C99-snprintf module we're using does just that.

For now, I think this is good enough.