FMI-2180 Copy memory issue fixes from dev_3.0 branch
This should make a majority of the memory leaks go away. There are still
2 memory leaks:

1. From the "ctest_fmi1_logger_test_check" test that compares the
   outputs of logging from previous test. This is test-only and will be
   fixed when merging 3.0 branch.

2. fmi_import_get_fmi_version: this is not fixed on 3.0 branch and I'll
   do it in next commit.

Memcheck tests are only run locally here, but on Jenkins in dev_3.0
branch. Did not include that since it will cause conflicts and pain when
later merging.

NOTE:
This commit should be reverted before merging dev_3.0 to avoid
conflicts.